LayoutTests/imported/w3c/web-platform-tests/pointerevents/mouse-pointer-boundary-events-for-shadowdom-expected.txt

PointerEvent: Verifies that mouse boundary events don't point to shadow-dom



FAIL PointerEvent: Verifies that mouse boundary events don't point to shadow-dom assert_array_equals: Moved into <input> expected property 1 to be "pointerover target=INPUT relatedTarget=BODY" but got "mouseout target=BODY relatedTarget=INPUT" (expected array ["pointerout target=BODY relatedTarget=INPUT", "pointerover target=INPUT relatedTarget=BODY", "mouseout target=BODY relatedTarget=INPUT", "mouseover target=INPUT relatedTarget=BODY"] got ["pointerout target=BODY relatedTarget=INPUT", "mouseout target=BODY relatedTarget=INPUT", "pointerover target=INPUT relatedTarget=BODY", "mouseover target=INPUT relatedTarget=BODY"])
